refactor: 클래스명 구체화
