+++ src/main/java/lotto/controller/WinningNumberGenerationController.java
@@ -2,25 +2,25 @@ package lotto.controller;
 
 import lotto.model.Lotto;
 import lotto.temp.IoController;
-import lotto.temp.Winning;
+import lotto.model.WinningNumberGenerator;
 import lotto.util.CommonIo;
 import lotto.view.InputView;
 
-public class WinningController {
-    private final Winning winning;
+public class WinningNumberGenerationController {
+    private final WinningNumberGenerator winningNumberGenerator;
     private final InputView inputView;
     private final IoController ioController;
 
-    public WinningController(Winning winning) {
+    public WinningNumberGenerationController(WinningNumberGenerator winningNumberGenerator) {
         CommonIo commonIo = new CommonIo();
         this.inputView = new InputView(commonIo);
         this.ioController = new IoController(commonIo);
-        this.winning = winning;
+        this.winningNumberGenerator = winningNumberGenerator;
     }
 
     public Lotto createWinningNumber(){
         inputView.printRequestWinningNumbers();
         String rawWinningNumbers = ioController.inputWinningNumbers();
-        return winning.createWinningNumbers(rawWinningNumbers);
+        return winningNumberGenerator.createWinningNumbers(rawWinningNumbers);
     }
 }
